% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasFIA-package.R
\docType{package}
\name{plasFIA-package}
\alias{plasFIA-package}
\alias{plasFIA}
\title{FIA-HRMS spiked plasma dataset.}
\description{
6 samples of human plasma spiked with 3 different concentrations
of molecules ranging from 10 ng/mL to 1000ng/mL in 2 replicates 
have been acquired on an Orbitrap Fusion. The resulting data are 
in the mzML directory under mzML centroid format, the \code{proFIAset}
object containg the integrated signal is in the \code{plasSet} object
information on the samples is in \code{plasSamples} and informations
about molecules is in \code{plasMols}.
}
