% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasFIA-package.R
\docType{data}
\name{plasMols}
\alias{plasMols}
\title{Information on the 40 molecules from the plasFIA package.}
\format{A data frame with 40 rows and 5 variables}
\usage{
data(plasMols)
}
\description{
A dataset giving the formula, the name, the class, the mass 
and the mass of the single-charged molecule of the plasFIA dataset.
\itemize{
  \item formula. The formula of the molecule
  \item names. The usual names of the molecules
  \item classes. The chemical family of the molecules.
  \item mass. The mass of the molecules.
  \item mass_M+H. The mass of the signly charge ion.
  ...
}
}
\keyword{datasets}
