% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasFIA-package.R
\docType{data}
\name{plasSet}
\alias{plasSet}
\title{Signal detected by the proFIA package in the plasFIA dataset.}
\format{A "proFIAset" object containing  1  classes.
Missing values have been imputed. 
2557  peaks detected.
629  features have been grouped.
The data matrix is avalaible.
Memory usage: 0.428 MB}
\usage{
data(plasSet)
}
\description{
A \code{proFIAset} object giving the signals detected
in the mzML files of the plasFIA datasets using the 
\code{analyzeAcquisitionFIA} function with the following parameters :
\itemize{
  \item ppm. 2
  \item fracGroup. 0.2
  \item ppmgroup. 0.5
  \item k. 2
}
}
\details{
Which were adapted to an Orbitrap Fusion.
}
\keyword{dataset}

